--- frontend/app/client/_app/immutable/chunks/Database.7f6fed06.js
@@ -1,1 +1,0 @@
-import{p as Y}from"./parse.bee59afc.js";import{j as J}from"./singletons.9380a151.js";import{i as V}from"./navigation.5a144dea.js";function z(r){return J.apply_action(r)}function Q(r){const o=JSON.parse(r);return o.data&&(o.data=Y(o.data)),o}function $(r){return HTMLElement.prototype.cloneNode.call(r)}function wt(r,o=()=>{}){const t=async({action:n,result:i,reset:s=!0,invalidateAll:a=!0})=>{i.type==="success"&&(s&&HTMLFormElement.prototype.reset.call(r),a&&await V()),(location.origin+location.pathname===n.origin+n.pathname||i.type==="redirect"||i.type==="error")&&z(i)};async function e(n){var v,C,g,I;if(((v=n.submitter)!=null&&v.hasAttribute("formmethod")?n.submitter.formMethod:$(r).method)!=="post")return;n.preventDefault();const s=new URL((C=n.submitter)!=null&&C.hasAttribute("formaction")?n.submitter.formAction:$(r).action),a=new FormData(r),u=(g=n.submitter)==null?void 0:g.getAttribute("name");u&&a.append(u,((I=n.submitter)==null?void 0:I.getAttribute("value"))??"");const l=new AbortController;let d=!1;const h=await o({action:s,cancel:()=>d=!0,controller:l,get data(){return a},formData:a,get form(){return r},formElement:r,submitter:n.submitter})??t;if(d)return;let f;try{const p=await fetch(s,{method:"POST",headers:{accept:"application/json","x-sveltekit-action":"true"},cache:"no-store",body:a,signal:l.signal});f=Q(await p.text()),f.type==="error"&&(f.status=p.status)}catch(p){if((p==null?void 0:p.name)==="AbortError")return;f={type:"error",error:p}}h({action:s,get data(){return a},formData:a,get form(){return r},formElement:r,update:p=>t({action:s,result:f,reset:p==null?void 0:p.reset,invalidateAll:p==null?void 0:p.invalidateAll}),result:f})}return HTMLFormElement.prototype.addEventListener.call(r,"submit",e),{destroy(){HTMLFormElement.prototype.removeEventListener.call(r,"submit",e)}}}var x=function(r,o){return x=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(t,e){t.__proto__=e}||function(t,e){for(var n in e)Object.prototype.hasOwnProperty.call(e,n)&&(t[n]=e[n])},x(r,o)};function y(r,o){if(typeof o!="function"&&o!==null)throw new TypeError("Class extends value "+String(o)+" is not a constructor or null");function t(){this.constructor=r}x(r,o),r.prototype=o===null?Object.create(o):(t.prototype=o.prototype,new t)}var M=function(){return M=Object.assign||function(o){for(var t,e=1,n=arguments.length;e<n;e++)for(var i in t=arguments[e])Object.prototype.hasOwnProperty.call(t,i)&&(o[i]=t[i]);return o},M.apply(this,arguments)};function m(r,o,t,e){return new(t||(t=Promise))(function(n,i){function s(l){try{u(e.next(l))}catch(d){i(d)}}function a(l){try{u(e.throw(l))}catch(d){i(d)}}function u(l){l.done?n(l.value):function(c){return c instanceof t?c:new t(function(h){h(c)})}(l.value).then(s,a)}u((e=e.apply(r,o||[])).next())})}function b(r,o){var t,e,n,i,s={label:0,sent:function(){if(1&n[0])throw n[1];return n[1]},trys:[],ops:[]};return i={next:a(0),throw:a(1),return:a(2)},typeof Symbol=="function"&&(i[Symbol.iterator]=function(){return this}),i;function a(u){return function(l){return function(c){if(t)throw new TypeError("Generator is already executing.");for(;i&&(i=0,c[0]&&(s=0)),s;)try{if(t=1,e&&(n=2&c[0]?e.return:c[0]?e.throw||((n=e.return)&&n.call(e),0):e.next)&&!(n=n.call(e,c[1])).done)return n;switch(e=0,n&&(c=[2&c[0],n.value]),c[0]){case 0:case 1:n=c;break;case 4:return s.label++,{value:c[1],done:!1};case 5:s.label++,e=c[1],c=[0];continue;case 7:c=s.ops.pop(),s.trys.pop();continue;default:if(n=s.trys,!((n=n.length>0&&n[n.length-1])||c[0]!==6&&c[0]!==2)){s=0;continue}if(c[0]===3&&(!n||c[1]>n[0]&&c[1]<n[3])){s.label=c[1];break}if(c[0]===6&&s.label<n[1]){s.label=n[1],n=c;break}if(n&&s.label<n[2]){s.label=n[2],s.ops.push(c);break}n[2]&&s.ops.pop(),s.trys.pop();continue}c=o.call(r,s)}catch(h){c=[6,h],e=0}finally{t=n=0}if(5&c[0])throw c[1];return{value:c[0]?c[1]:void 0,done:!0}}([u,l])}}}var G,O=function(r){function o(t){var e,n,i,s,a=this;return(a=r.call(this,"ClientResponseError")||this).url="",a.status=0,a.response={},a.isAbort=!1,a.originalError=null,Object.setPrototypeOf(a,o.prototype),t!==null&&typeof t=="object"&&(a.url=typeof t.url=="string"?t.url:"",a.status=typeof t.status=="number"?t.status:0,a.isAbort=!!t.isAbort,a.originalError=t.originalError,t.response!==null&&typeof t.response=="object"?a.response=t.response:t.data!==null&&typeof t.data=="object"?a.response=t.data:a.response={}),a.originalError||t instanceof o||(a.originalError=t),typeof DOMException<"u"&&t instanceof DOMException&&(a.isAbort=!0),a.name="ClientResponseError "+a.status,a.message=(e=a.response)===null||e===void 0?void 0:e.message,a.message||(a.isAbort?a.message="The request was autocancelled. You can find more info in https://github.com/pocketbase/js-sdk#auto-cancellation.":!((s=(i=(n=a.originalError)===null||n===void 0?void 0:n.cause)===null||i===void 0?void 0:i.message)===null||s===void 0)&&s.includes("ECONNREFUSED ::1")?a.message="Failed to connect to the PocketBase server. Try changing the SDK URL from localhost to 127.0.0.1 (https://github.com/pocketbase/js-sdk/issues/21).":a.message="Something went wrong while processing your request."),a}return y(o,r),Object.defineProperty(o.prototype,"data",{get:function(){return this.response},enumerable:!1,configurable:!0}),o.prototype.toJSON=function(){return M({},this)},o}(Error),N=/^[\u0009\u0020-\u007e\u0080-\u00ff]+$/;function B(r,o,t){var e=Object.assign({},t||{}),n=e.encode||Z;if(!N.test(r))throw new TypeError("argument name is invalid");var i=n(o);if(i&&!N.test(i))throw new TypeError("argument val is invalid");var s=r+"="+i;if(e.maxAge!=null){var a=e.maxAge-0;if(isNaN(a)||!isFinite(a))throw new TypeError("option maxAge is invalid");s+="; Max-Age="+Math.floor(a)}if(e.domain){if(!N.test(e.domain))throw new TypeError("option domain is invalid");s+="; Domain="+e.domain}if(e.path){if(!N.test(e.path))throw new TypeError("option path is invalid");s+="; Path="+e.path}if(e.expires){if(!function(l){return Object.prototype.toString.call(l)==="[object Date]"||l instanceof Date}(e.expires)||isNaN(e.expires.valueOf()))throw new TypeError("option expires is invalid");s+="; Expires="+e.expires.toUTCString()}if(e.httpOnly&&(s+="; HttpOnly"),e.secure&&(s+="; Secure"),e.priority)switch(typeof e.priority=="string"?e.priority.toLowerCase():e.priority){case"low":s+="; Priority=Low";break;case"medium":s+="; Priority=Medium";break;case"high":s+="; Priority=High";break;default:throw new TypeError("option priority is invalid")}if(e.sameSite)switch(typeof e.sameSite=="string"?e.sameSite.toLowerCase():e.sameSite){case!0:s+="; SameSite=Strict";break;case"lax":s+="; SameSite=Lax";break;case"strict":s+="; SameSite=Strict";break;case"none":s+="; SameSite=None";break;default:throw new TypeError("option sameSite is invalid")}return s}function X(r){return r.indexOf("%")!==-1?decodeURIComponent(r):r}function Z(r){return encodeURIComponent(r)}function H(r){if(r)try{var o=decodeURIComponent(G(r.split(".")[1]).split("").map(function(t){return"%"+("00"+t.charCodeAt(0).toString(16)).slice(-2)}).join(""));return JSON.parse(o)||{}}catch{}return{}}function tt(r,o){o===void 0&&(o=0);var t=H(r);return!(Object.keys(t).length>0&&(!t.exp||t.exp-o>Date.now()/1e3))}G=typeof atob=="function"?atob:function(r){var o=String(r).replace(/=+$/,"");if(o.length%4==1)throw new Error("'atob' failed: The string to be decoded is not correctly encoded.");for(var t,e,n=0,i=0,s="";e=o.charAt(i++);~e&&(t=n%4?64*t+e:e,n++%4)?s+=String.fromCharCode(255&t>>(-2*n&6)):0)e="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=".indexOf(e);return s};var j=function(){function r(o){o===void 0&&(o={}),this.$load(o||{})}return r.prototype.load=function(o){return this.$load(o)},r.prototype.$load=function(o){for(var t=0,e=Object.entries(o);t<e.length;t++){var n=e[t],i=n[0],s=n[1];this[i]=s}this.id=o.id!==void 0?o.id:"",this.created=o.created!==void 0?o.created:"",this.updated=o.updated!==void 0?o.updated:""},Object.defineProperty(r.prototype,"$isNew",{get:function(){return!this.id},enumerable:!1,configurable:!0}),r.prototype.clone=function(){return this.$clone()},r.prototype.$clone=function(){var o=typeof structuredClone=="function"?structuredClone(this):JSON.parse(JSON.stringify(this));return new this.constructor(o)},r.prototype.export=function(){return this.$export()},r.prototype.$export=function(){return typeof structuredClone=="function"?structuredClone(this):Object.assign({},this)},r}(),U=function(r){function o(){return r!==null&&r.apply(this,arguments)||this}return y(o,r),o.prototype.$load=function(t){r.prototype.$load.call(this,t),this.collectionId=typeof t.collectionId=="string"?t.collectionId:"",this.collectionName=typeof t.collectionName=="string"?t.collectionName:"",this._loadExpand(t.expand)},o.prototype._loadExpand=function(t){for(var e in t=t||{},this.expand={},t)Array.isArray(t[e])?this.expand[e]=t[e].map(function(n){return new o(n||{})}):this.expand[e]=new o(t[e]||{})},o}(j),D=function(r){function o(){return r!==null&&r.apply(this,arguments)||this}return y(o,r),o.prototype.$load=function(t){r.prototype.$load.call(this,t),this.avatar=typeof t.avatar=="number"?t.avatar:0,this.email=typeof t.email=="string"?t.email:""},o}(j),q="pb_auth",et=function(){function r(){this.baseToken="",this.baseModel=null,this._onChangeCallbacks=[]}return Object.defineProperty(r.prototype,"token",{get:function(){return this.baseToken},enumerable:!1,configurable:!0}),Object.defineProperty(r.prototype,"model",{get:function(){return this.baseModel},enumerable:!1,configurable:!0}),Object.defineProperty(r.prototype,"isValid",{get:function(){return!tt(this.token)},enumerable:!1,configurable:!0}),r.prototype.save=function(o,t){this.baseToken=o||"",this.baseModel=t!==null&&typeof t=="object"?t.collectionId!==void 0?new U(t):new D(t):null,this.triggerChange()},r.prototype.clear=function(){this.baseToken="",this.baseModel=null,this.triggerChange()},r.prototype.loadFromCookie=function(o,t){t===void 0&&(t=q);var e=function(s,a){var u={};if(typeof s!="string")return u;for(var l=Object.assign({},a||{}).decode||X,d=0;d<s.length;){var c=s.indexOf("=",d);if(c===-1)break;var h=s.indexOf(";",d);if(h===-1)h=s.length;else if(h<c){d=s.lastIndexOf(";",c-1)+1;continue}var f=s.slice(d,c).trim();if(u[f]===void 0){var v=s.slice(c+1,h).trim();v.charCodeAt(0)===34&&(v=v.slice(1,-1));try{u[f]=l(v)}catch{u[f]=v}}d=h+1}return u}(o||"")[t]||"",n={};try{(typeof(n=JSON.parse(e))===null||typeof n!="object"||Array.isArray(n))&&(n={})}catch{}this.save(n.token||"",n.model||null)},r.prototype.exportToCookie=function(o,t){var e,n,i;t===void 0&&(t=q);var s={secure:!0,sameSite:!0,httpOnly:!0,path:"/"},a=H(this.token);a!=null&&a.exp?s.expires=new Date(1e3*a.exp):s.expires=new Date("1970-01-01"),o=Object.assign({},s,o);var u={token:this.token,model:((e=this.model)===null||e===void 0?void 0:e.export())||null},l=B(t,JSON.stringify(u),o),d=typeof Blob<"u"?new Blob([l]).size:l.length;return u.model&&d>4096&&(u.model={id:(n=u==null?void 0:u.model)===null||n===void 0?void 0:n.id,email:(i=u==null?void 0:u.model)===null||i===void 0?void 0:i.email},this.model instanceof U&&(u.model.username=this.model.username,u.model.verified=this.model.verified,u.model.collectionId=this.model.collectionId),l=B(t,JSON.stringify(u),o)),l},r.prototype.onChange=function(o,t){var e=this;return t===void 0&&(t=!1),this._onChangeCallbacks.push(o),t&&o(this.token,this.model),function(){for(var n=e._onChangeCallbacks.length-1;n>=0;n--)if(e._onChangeCallbacks[n]==o)return delete e._onChangeCallbacks[n],void e._onChangeCallbacks.splice(n,1)}},r.prototype.triggerChange=function(){for(var o=0,t=this._onChangeCallbacks;o<t.length;o++){var e=t[o];e&&e(this.token,this.model)}},r}(),nt=function(r){function o(t){t===void 0&&(t="pocketbase_auth");var e=r.call(this)||this;return e.storageFallback={},e.storageKey=t,e}return y(o,r),Object.defineProperty(o.prototype,"token",{get:function(){return(this._storageGet(this.storageKey)||{}).token||""},enumerable:!1,configurable:!0}),Object.defineProperty(o.prototype,"model",{get:function(){var t,e=this._storageGet(this.storageKey)||{};return e===null||typeof e!="object"||e.model===null||typeof e.model!="object"?null:((t=e.model)===null||t===void 0?void 0:t.collectionId)===void 0?new D(e.model):new U(e.model)},enumerable:!1,configurable:!0}),o.prototype.save=function(t,e){this._storageSet(this.storageKey,{token:t,model:e}),r.prototype.save.call(this,t,e)},o.prototype.clear=function(){this._storageRemove(this.storageKey),r.prototype.clear.call(this)},o.prototype._storageGet=function(t){if(typeof window<"u"&&(window!=null&&window.localStorage)){var e=window.localStorage.getItem(t)||"";try{return JSON.parse(e)}catch{return e}}return this.storageFallback[t]},o.prototype._storageSet=function(t,e){if(typeof window<"u"&&(window!=null&&window.localStorage)){var n=e;typeof e!="string"&&(n=JSON.stringify(e)),window.localStorage.setItem(t,n)}else this.storageFallback[t]=e},o.prototype._storageRemove=function(t){var e;typeof window<"u"&&(window!=null&&window.localStorage)&&((e=window.localStorage)===null||e===void 0||e.removeItem(t)),delete this.storageFallback[t]},o}(et),_=function(o){this.client=o},ot=function(r){function o(){return r!==null&&r.apply(this,arguments)||this}return y(o,r),o.prototype.getAll=function(t){return t===void 0&&(t={}),this.client.send("/api/settings",{method:"GET",params:t}).then(function(e){return e||{}})},o.prototype.update=function(t,e){return t===void 0&&(t={}),e===void 0&&(e={}),this.client.send("/api/settings",{method:"PATCH",params:e,body:t}).then(function(n){return n||{}})},o.prototype.testS3=function(t,e){t===void 0&&(t="storage"),e===void 0&&(e={});var n={filesystem:t};return this.client.send("/api/settings/test/s3",{method:"POST",params:e,body:n}).then(function(){return!0})},o.prototype.testEmail=function(t,e,n){n===void 0&&(n={});var i={email:t,template:e};return this.client.send("/api/settings/test/email",{method:"POST",params:n,body:i}).then(function(){return!0})},o.prototype.generateAppleClientSecret=function(t,e,n,i,s,a,u){return a===void 0&&(a={}),u===void 0&&(u={}),a=Object.assign({clientId:t,teamId:e,keyId:n,privateKey:i,duration:s},a),this.client.send("/api/settings/apple/generate-client-secret",{method:"POST",params:u,body:a})},o}(_),K=function(o,t,e,n,i){this.page=o>0?o:1,this.perPage=t>=0?t:0,this.totalItems=e>=0?e:0,this.totalPages=n>=0?n:0,this.items=i||[]},F=function(r){function o(){return r!==null&&r.apply(this,arguments)||this}return y(o,r),o.prototype.getFullList=function(t,e){if(typeof t=="number")return this._getFullList(this.baseCrudPath,t,e);var n=Object.assign({},t,e);return this._getFullList(this.baseCrudPath,n.batch||200,n)},o.prototype.getList=function(t,e,n){return t===void 0&&(t=1),e===void 0&&(e=30),n===void 0&&(n={}),this._getList(this.baseCrudPath,t,e,n)},o.prototype.getFirstListItem=function(t,e){return e===void 0&&(e={}),this._getFirstListItem(this.baseCrudPath,t,e)},o.prototype.getOne=function(t,e){return e===void 0&&(e={}),this._getOne(this.baseCrudPath,t,e)},o.prototype.create=function(t,e){return t===void 0&&(t={}),e===void 0&&(e={}),this._create(this.baseCrudPath,t,e)},o.prototype.update=function(t,e,n){return e===void 0&&(e={}),n===void 0&&(n={}),this._update(this.baseCrudPath,t,e,n)},o.prototype.delete=function(t,e){return e===void 0&&(e={}),this._delete(this.baseCrudPath,t,e)},o}(function(r){function o(){return r!==null&&r.apply(this,arguments)||this}return y(o,r),o.prototype._getFullList=function(t,e,n){var i=this;e===void 0&&(e=200),n===void 0&&(n={});var s=[],a=function(u){return m(i,void 0,void 0,function(){return b(this,function(l){return[2,this._getList(t,u,e||200,n).then(function(d){var c=d,h=c.items,f=c.totalItems;return s=s.concat(h),h.length&&f>s.length?a(u+1):s})]})})};return a(1)},o.prototype._getList=function(t,e,n,i){var s=this;return e===void 0&&(e=1),n===void 0&&(n=30),i===void 0&&(i={}),i=Object.assign({page:e,perPage:n},i),this.client.send(t,{method:"GET",params:i}).then(function(a){var u=[];if(a!=null&&a.items){a.items=a.items||[];for(var l=0,d=a.items;l<d.length;l++){var c=d[l];u.push(s.decode(c))}}return new K((a==null?void 0:a.page)||1,(a==null?void 0:a.perPage)||0,(a==null?void 0:a.totalItems)||0,(a==null?void 0:a.totalPages)||0,u)})},o.prototype._getOne=function(t,e,n){var i=this;return n===void 0&&(n={}),this.client.send(t+"/"+encodeURIComponent(e),{method:"GET",params:n}).then(function(s){return i.decode(s)})},o.prototype._getFirstListItem=function(t,e,n){return n===void 0&&(n={}),n=Object.assign({filter:e,$cancelKey:"one_by_filter_"+t+"_"+e},n),this._getList(t,1,1,n).then(function(i){var s;if(!(!((s=i==null?void 0:i.items)===null||s===void 0)&&s.length))throw new O({status:404,data:{code:404,message:"The requested resource wasn't found.",data:{}}});return i.items[0]})},o.prototype._create=function(t,e,n){var i=this;return e===void 0&&(e={}),n===void 0&&(n={}),this.client.send(t,{method:"POST",params:n,body:e}).then(function(s){return i.decode(s)})},o.prototype._update=function(t,e,n,i){var s=this;return n===void 0&&(n={}),i===void 0&&(i={}),this.client.send(t+"/"+encodeURIComponent(e),{method:"PATCH",params:i,body:n}).then(function(a){return s.decode(a)})},o.prototype._delete=function(t,e,n){return n===void 0&&(n={}),this.client.send(t+"/"+encodeURIComponent(e),{method:"DELETE",params:n}).then(function(){return!0})},o}(_)),it=function(r){function o(){return r!==null&&r.apply(this,arguments)||this}return y(o,r),o.prototype.decode=function(t){return new D(t)},Object.defineProperty(o.prototype,"baseCrudPath",{get:function(){return"/api/admins"},enumerable:!1,configurable:!0}),o.prototype.update=function(t,e,n){var i=this;return e===void 0&&(e={}),n===void 0&&(n={}),r.prototype.update.call(this,t,e,n).then(function(s){var a,u;return i.client.authStore.model&&((a=i.client.authStore.model)===null||a===void 0?void 0:a.collectionId)===void 0&&((u=i.client.authStore.model)===null||u===void 0?void 0:u.id)===(s==null?void 0:s.id)&&i.client.authStore.save(i.client.authStore.token,s),s})},o.prototype.delete=function(t,e){var n=this;return e===void 0&&(e={}),r.prototype.delete.call(this,t,e).then(function(i){var s,a;return i&&n.client.authStore.model&&((s=n.client.authStore.model)===null||s===void 0?void 0:s.collectionId)===void 0&&((a=n.client.authStore.model)===null||a===void 0?void 0:a.id)===t&&n.client.authStore.clear(),i})},o.prototype.authResponse=function(t){var e=this.decode((t==null?void 0:t.admin)||{});return t!=null&&t.token&&(t!=null&&t.admin)&&this.client.authStore.save(t.token,e),Object.assign({},t,{token:(t==null?void 0:t.token)||"",admin:e})},o.prototype.authWithPassword=function(t,e,n,i){return n===void 0&&(n={}),i===void 0&&(i={}),n=Object.assign({identity:t,password:e},n),this.client.send(this.baseCrudPath+"/auth-with-password",{method:"POST",params:i,body:n}).then(this.authResponse.bind(this))},o.prototype.authRefresh=function(t,e){return t===void 0&&(t={}),e===void 0&&(e={}),this.client.send(this.baseCrudPath+"/auth-refresh",{method:"POST",params:e,body:t}).then(this.authResponse.bind(this))},o.prototype.requestPasswordReset=function(t,e,n){return e===void 0&&(e={}),n===void 0&&(n={}),e=Object.assign({email:t},e),this.client.send(this.baseCrudPath+"/request-password-reset",{method:"POST",params:n,body:e}).then(function(){return!0})},o.prototype.confirmPasswordReset=function(t,e,n,i,s){return i===void 0&&(i={}),s===void 0&&(s={}),i=Object.assign({token:t,password:e,passwordConfirm:n},i),this.client.send(this.baseCrudPath+"/confirm-password-reset",{method:"POST",params:s,body:i}).then(function(){return!0})},o}(F),rt=function(r){function o(){return r!==null&&r.apply(this,arguments)||this}return y(o,r),o.prototype.$load=function(t){r.prototype.$load.call(this,t),this.recordId=typeof t.recordId=="string"?t.recordId:"",this.collectionId=typeof t.collectionId=="string"?t.collectionId:"",this.provider=typeof t.provider=="string"?t.provider:"",this.providerId=typeof t.providerId=="string"?t.providerId:""},o}(j),st=function(r){function o(t,e){var n=r.call(this,t)||this;return n.collectionIdOrName=e,n}return y(o,r),o.prototype.decode=function(t){return new U(t)},Object.defineProperty(o.prototype,"baseCrudPath",{get:function(){return this.baseCollectionPath+"/records"},enumerable:!1,configurable:!0}),Object.defineProperty(o.prototype,"baseCollectionPath",{get:function(){return"/api/collections/"+encodeURIComponent(this.collectionIdOrName)},enumerable:!1,configurable:!0}),o.prototype.subscribeOne=function(t,e){return m(this,void 0,void 0,function(){return b(this,function(n){return console.warn("PocketBase: subscribeOne(recordId, callback) is deprecated. Please replace it with subscribe(recordId, callback)."),[2,this.client.realtime.subscribe(this.collectionIdOrName+"/"+t,e)]})})},o.prototype.subscribe=function(t,e){return m(this,void 0,void 0,function(){var n;return b(this,function(i){if(typeof t=="function")return console.warn("PocketBase: subscribe(callback) is deprecated. Please replace it with subscribe('*', callback)."),[2,this.client.realtime.subscribe(this.collectionIdOrName,t)];if(!e)throw new Error("Missing subscription callback.");if(t==="")throw new Error("Missing topic.");return n=this.collectionIdOrName,t!=="*"&&(n+="/"+t),[2,this.client.realtime.subscribe(n,e)]})})},o.prototype.unsubscribe=function(t){return m(this,void 0,void 0,function(){return b(this,function(e){return t==="*"?[2,this.client.realtime.unsubscribe(this.collectionIdOrName)]:t?[2,this.client.realtime.unsubscribe(this.collectionIdOrName+"/"+t)]:[2,this.client.realtime.unsubscribeByPrefix(this.collectionIdOrName)]})})},o.prototype.getFullList=function(t,e){if(typeof t=="number")return r.prototype.getFullList.call(this,t,e);var n=Object.assign({},t,e);return r.prototype.getFullList.call(this,n)},o.prototype.getList=function(t,e,n){return t===void 0&&(t=1),e===void 0&&(e=30),n===void 0&&(n={}),r.prototype.getList.call(this,t,e,n)},o.prototype.getFirstListItem=function(t,e){return e===void 0&&(e={}),r.prototype.getFirstListItem.call(this,t,e)},o.prototype.getOne=function(t,e){return e===void 0&&(e={}),r.prototype.getOne.call(this,t,e)},o.prototype.create=function(t,e){return t===void 0&&(t={}),e===void 0&&(e={}),r.prototype.create.call(this,t,e)},o.prototype.update=function(t,e,n){var i=this;return e===void 0&&(e={}),n===void 0&&(n={}),r.prototype.update.call(this,t,e,n).then(function(s){var a,u,l;return((a=i.client.authStore.model)===null||a===void 0?void 0:a.id)!==(s==null?void 0:s.id)||((u=i.client.authStore.model)===null||u===void 0?void 0:u.collectionId)!==i.collectionIdOrName&&((l=i.client.authStore.model)===null||l===void 0?void 0:l.collectionName)!==i.collectionIdOrName||i.client.authStore.save(i.client.authStore.token,s),s})},o.prototype.delete=function(t,e){var n=this;return e===void 0&&(e={}),r.prototype.delete.call(this,t,e).then(function(i){var s,a,u;return!i||((s=n.client.authStore.model)===null||s===void 0?void 0:s.id)!==t||((a=n.client.authStore.model)===null||a===void 0?void 0:a.collectionId)!==n.collectionIdOrName&&((u=n.client.authStore.model)===null||u===void 0?void 0:u.collectionName)!==n.collectionIdOrName||n.client.authStore.clear(),i})},o.prototype.authResponse=function(t){var e=this.decode((t==null?void 0:t.record)||{});return this.client.authStore.save(t==null?void 0:t.token,e),Object.assign({},t,{token:(t==null?void 0:t.token)||"",record:e})},o.prototype.listAuthMethods=function(t){return t===void 0&&(t={}),this.client.send(this.baseCollectionPath+"/auth-methods",{method:"GET",params:t}).then(function(e){return Object.assign({},e,{usernamePassword:!!(e!=null&&e.usernamePassword),emailPassword:!!(e!=null&&e.emailPassword),authProviders:Array.isArray(e==null?void 0:e.authProviders)?e==null?void 0:e.authProviders:[]})})},o.prototype.authWithPassword=function(t,e,n,i){var s=this;return n===void 0&&(n={}),i===void 0&&(i={}),n=Object.assign({identity:t,password:e},n),this.client.send(this.baseCollectionPath+"/auth-with-password",{method:"POST",params:i,body:n}).then(function(a){return s.authResponse(a)})},o.prototype.authWithOAuth2Code=function(t,e,n,i,s,a,u){var l=this;return s===void 0&&(s={}),a===void 0&&(a={}),u===void 0&&(u={}),a=Object.assign({provider:t,code:e,codeVerifier:n,redirectUrl:i,createData:s},a),this.client.send(this.baseCollectionPath+"/auth-with-oauth2",{method:"POST",params:u,body:a}).then(function(d){return l.authResponse(d)})},o.prototype.authWithOAuth2=function(){for(var t=[],e=0;e<arguments.length;e++)t[e]=arguments[e];return m(this,void 0,void 0,function(){var n,i,s,a,u=this;return b(this,function(l){switch(l.label){case 0:return t.length>1||typeof(t==null?void 0:t[0])=="string"?(console.warn("PocketBase: This form of authWithOAuth2() is deprecated and may get removed in the future. Please replace with authWithOAuth2Code() OR use the authWithOAuth2() realtime form as shown in https://pocketbase.io/docs/authentication/#oauth2-integration."),[2,this.authWithOAuth2Code((t==null?void 0:t[0])||"",(t==null?void 0:t[1])||"",(t==null?void 0:t[2])||"",(t==null?void 0:t[3])||"",(t==null?void 0:t[4])||{},(t==null?void 0:t[5])||{},(t==null?void 0:t[6])||{})]):(n=(t==null?void 0:t[0])||{},[4,this.listAuthMethods()]);case 1:if(i=l.sent(),!(s=i.authProviders.find(function(d){return d.name===n.provider})))throw new O(new Error('Missing or invalid provider "'.concat(n.provider,'".')));return a=this.client.buildUrl("/api/oauth2-redirect"),[2,new Promise(function(d,c){return m(u,void 0,void 0,function(){var h,f,v,C,g,I=this;return b(this,function(p){switch(p.label){case 0:return p.trys.push([0,3,,4]),[4,this.client.realtime.subscribe("@oauth2",function(T){return m(I,void 0,void 0,function(){var L,k,S;return b(this,function(w){switch(w.label){case 0:L=this.client.realtime.clientId,w.label=1;case 1:if(w.trys.push([1,3,,4]),h(),!T.state||L!==T.state)throw new Error("State parameters don't match.");return[4,this.authWithOAuth2Code(s.name,T.code,s.codeVerifier,a,n.createData,n.body,n.query)];case 2:return k=w.sent(),d(k),[3,4];case 3:return S=w.sent(),c(new O(S)),[3,4];case 4:return[2]}})})})];case 1:return h=p.sent(),f={state:this.client.realtime.clientId},!((g=n.scopes)===null||g===void 0)&&g.length&&(f.scope=n.scopes.join(" ")),v=this._replaceQueryParams(s.authUrl+a,f),[4,n.urlCallback?n.urlCallback(v):this._defaultUrlCallback(v)];case 2:return p.sent(),[3,4];case 3:return C=p.sent(),c(new O(C)),[3,4];case 4:return[2]}})})})]}})})},o.prototype.authRefresh=function(t,e){var n=this;return t===void 0&&(t={}),e===void 0&&(e={}),this.client.send(this.baseCollectionPath+"/auth-refresh",{method:"POST",params:e,body:t}).then(function(i){return n.authResponse(i)})},o.prototype.requestPasswordReset=function(t,e,n){return e===void 0&&(e={}),n===void 0&&(n={}),e=Object.assign({email:t},e),this.client.send(this.baseCollectionPath+"/request-password-reset",{method:"POST",params:n,body:e}).then(function(){return!0})},o.prototype.confirmPasswordReset=function(t,e,n,i,s){return i===void 0&&(i={}),s===void 0&&(s={}),i=Object.assign({token:t,password:e,passwordConfirm:n},i),this.client.send(this.baseCollectionPath+"/confirm-password-reset",{method:"POST",params:s,body:i}).then(function(){return!0})},o.prototype.requestVerification=function(t,e,n){return e===void 0&&(e={}),n===void 0&&(n={}),e=Object.assign({email:t},e),this.client.send(this.baseCollectionPath+"/request-verification",{method:"POST",params:n,body:e}).then(function(){return!0})},o.prototype.confirmVerification=function(t,e,n){return e===void 0&&(e={}),n===void 0&&(n={}),e=Object.assign({token:t},e),this.client.send(this.baseCollectionPath+"/confirm-verification",{method:"POST",params:n,body:e}).then(function(){return!0})},o.prototype.requestEmailChange=function(t,e,n){return e===void 0&&(e={}),n===void 0&&(n={}),e=Object.assign({newEmail:t},e),this.client.send(this.baseCollectionPath+"/request-email-change",{method:"POST",params:n,body:e}).then(function(){return!0})},o.prototype.confirmEmailChange=function(t,e,n,i){return n===void 0&&(n={}),i===void 0&&(i={}),n=Object.assign({token:t,password:e},n),this.client.send(this.baseCollectionPath+"/confirm-email-change",{method:"POST",params:i,body:n}).then(function(){return!0})},o.prototype.listExternalAuths=function(t,e){return e===void 0&&(e={}),this.client.send(this.baseCrudPath+"/"+encodeURIComponent(t)+"/external-auths",{method:"GET",params:e}).then(function(n){var i=[];if(Array.isArray(n))for(var s=0,a=n;s<a.length;s++){var u=a[s];i.push(new rt(u))}return i})},o.prototype.unlinkExternalAuth=function(t,e,n){return n===void 0&&(n={}),this.client.send(this.baseCrudPath+"/"+encodeURIComponent(t)+"/external-auths/"+encodeURIComponent(e),{method:"DELETE",params:n}).then(function(){return!0})},o.prototype._replaceQueryParams=function(t,e){e===void 0&&(e={});var n=t,i="";t.indexOf("?")>=0&&(n=t.substring(0,t.indexOf("?")),i=t.substring(t.indexOf("?")+1));for(var s={},a=0,u=i.split("&");a<u.length;a++){var l=u[a];if(l!=""){var d=l.split("=");s[decodeURIComponent(d[0].replace(/\+/g," "))]=decodeURIComponent((d[1]||"").replace(/\+/g," "))}}for(var c in e)e.hasOwnProperty(c)&&(e[c]==null?delete s[c]:s[c]=e[c]);for(var c in i="",s)s.hasOwnProperty(c)&&(i!=""&&(i+="&"),i+=encodeURIComponent(c.replace(/%20/g,"+"))+"="+encodeURIComponent(s[c].replace(/%20/g,"+")));return i!=""?n+"?"+i:n},o.prototype._defaultUrlCallback=function(t){if(typeof window>"u"||!(window!=null&&window.open))throw new O(new Error("Not in a browser context - please pass a custom urlCallback function."));var e=1024,n=768,i=window.innerWidth,s=window.innerHeight,a=i/2-(e=e>i?i:e)/2,u=s/2-(n=n>s?s:n)/2;window.open(t,"oauth2-popup","width="+e+",height="+n+",top="+u+",left="+a+",resizable,menubar=no")},o}(F),at=function(o){o===void 0&&(o={}),this.id=o.id!==void 0?o.id:"",this.name=o.name!==void 0?o.name:"",this.type=o.type!==void 0?o.type:"text",this.system=!!o.system,this.required=!!o.required,this.options=typeof o.options=="object"&&o.options!==null?o.options:{}},ct=function(r){function o(){return r!==null&&r.apply(this,arguments)||this}return y(o,r),o.prototype.$load=function(t){r.prototype.$load.call(this,t),this.system=!!t.system,this.name=typeof t.name=="string"?t.name:"",this.type=typeof t.type=="string"?t.type:"base",this.options=t.options!==void 0&&t.options!==null?t.options:{},this.indexes=Array.isArray(t.indexes)?t.indexes:[],this.listRule=typeof t.listRule=="string"?t.listRule:null,this.viewRule=typeof t.viewRule=="string"?t.viewRule:null,this.createRule=typeof t.createRule=="string"?t.createRule:null,this.updateRule=typeof t.updateRule=="string"?t.updateRule:null,this.deleteRule=typeof t.deleteRule=="string"?t.deleteRule:null,t.schema=Array.isArray(t.schema)?t.schema:[],this.schema=[];for(var e=0,n=t.schema;e<n.length;e++){var i=n[e];this.schema.push(new at(i))}},Object.defineProperty(o.prototype,"isBase",{get:function(){return this.$isBase},enumerable:!1,configurable:!0}),Object.defineProperty(o.prototype,"$isBase",{get:function(){return this.type==="base"},enumerable:!1,configurable:!0}),Object.defineProperty(o.prototype,"isAuth",{get:function(){return this.$isAuth},enumerable:!1,configurable:!0}),Object.defineProperty(o.prototype,"$isAuth",{get:function(){return this.type==="auth"},enumerable:!1,configurable:!0}),Object.defineProperty(o.prototype,"isView",{get:function(){return this.$isView},enumerable:!1,configurable:!0}),Object.defineProperty(o.prototype,"$isView",{get:function(){return this.type==="view"},enumerable:!1,configurable:!0}),o}(j),ut=function(r){function o(){return r!==null&&r.apply(this,arguments)||this}return y(o,r),o.prototype.decode=function(t){return new ct(t)},Object.defineProperty(o.prototype,"baseCrudPath",{get:function(){return"/api/collections"},enumerable:!1,configurable:!0}),o.prototype.import=function(t,e,n){return e===void 0&&(e=!1),n===void 0&&(n={}),m(this,void 0,void 0,function(){return b(this,function(i){return[2,this.client.send(this.baseCrudPath+"/import",{method:"PUT",params:n,body:{collections:t,deleteMissing:e}}).then(function(){return!0})]})})},o}(F),W=function(r){function o(){return r!==null&&r.apply(this,arguments)||this}return y(o,r),o.prototype.$load=function(t){r.prototype.$load.call(this,t),t.remoteIp=t.remoteIp||t.ip,this.url=typeof t.url=="string"?t.url:"",this.method=typeof t.method=="string"?t.method:"GET",this.status=typeof t.status=="number"?t.status:200,this.auth=typeof t.auth=="string"?t.auth:"guest",this.remoteIp=typeof t.remoteIp=="string"?t.remoteIp:"",this.userIp=typeof t.userIp=="string"?t.userIp:"",this.referer=typeof t.referer=="string"?t.referer:"",this.userAgent=typeof t.userAgent=="string"?t.userAgent:"",this.meta=typeof t.meta=="object"&&t.meta!==null?t.meta:{}},o}(j),lt=function(r){function o(){return r!==null&&r.apply(this,arguments)||this}return y(o,r),o.prototype.getRequestsList=function(t,e,n){return t===void 0&&(t=1),e===void 0&&(e=30),n===void 0&&(n={}),n=Object.assign({page:t,perPage:e},n),this.client.send("/api/logs/requests",{method:"GET",params:n}).then(function(i){var s=[];if(i!=null&&i.items){i.items=(i==null?void 0:i.items)||[];for(var a=0,u=i.items;a<u.length;a++){var l=u[a];s.push(new W(l))}}return new K((i==null?void 0:i.page)||1,(i==null?void 0:i.perPage)||0,(i==null?void 0:i.totalItems)||0,(i==null?void 0:i.totalPages)||0,s)})},o.prototype.getRequest=function(t,e){return e===void 0&&(e={}),this.client.send("/api/logs/requests/"+encodeURIComponent(t),{method:"GET",params:e}).then(function(n){return new W(n)})},o.prototype.getRequestsStats=function(t){return t===void 0&&(t={}),this.client.send("/api/logs/requests/stats",{method:"GET",params:t}).then(function(e){return e})},o}(_),dt=function(r){function o(){var t=r!==null&&r.apply(this,arguments)||this;return t.clientId="",t.eventSource=null,t.subscriptions={},t.lastSentTopics=[],t.maxConnectTimeout=15e3,t.reconnectAttempts=0,t.maxReconnectAttempts=1/0,t.predefinedReconnectIntervals=[200,300,500,1e3,1200,1500,2e3],t.pendingConnects=[],t}return y(o,r),Object.defineProperty(o.prototype,"isConnected",{get:function(){return!!this.eventSource&&!!this.clientId&&!this.pendingConnects.length},enumerable:!1,configurable:!0}),o.prototype.subscribe=function(t,e){var n;return m(this,void 0,void 0,function(){var i,s=this;return b(this,function(a){switch(a.label){case 0:if(!t)throw new Error("topic must be set.");return i=function(u){var l,d=u;try{l=JSON.parse(d==null?void 0:d.data)}catch{}e(l||{})},this.subscriptions[t]||(this.subscriptions[t]=[]),this.subscriptions[t].push(i),this.isConnected?[3,2]:[4,this.connect()];case 1:return a.sent(),[3,5];case 2:return this.subscriptions[t].length!==1?[3,4]:[4,this.submitSubscriptions()];case 3:return a.sent(),[3,5];case 4:(n=this.eventSource)===null||n===void 0||n.addEventListener(t,i),a.label=5;case 5:return[2,function(){return m(s,void 0,void 0,function(){return b(this,function(u){return[2,this.unsubscribeByTopicAndListener(t,i)]})})}]}})})},o.prototype.unsubscribe=function(t){var e;return m(this,void 0,void 0,function(){var n,i,s;return b(this,function(a){switch(a.label){case 0:if(!this.hasSubscriptionListeners(t))return[2];if(t){for(n=0,i=this.subscriptions[t];n<i.length;n++)s=i[n],(e=this.eventSource)===null||e===void 0||e.removeEventListener(t,s);delete this.subscriptions[t]}else this.subscriptions={};return this.hasSubscriptionListeners()?[3,1]:(this.disconnect(),[3,3]);case 1:return this.hasSubscriptionListeners(t)?[3,3]:[4,this.submitSubscriptions()];case 2:a.sent(),a.label=3;case 3:return[2]}})})},o.prototype.unsubscribeByPrefix=function(t){var e;return m(this,void 0,void 0,function(){var n,i,s,a,u;return b(this,function(l){switch(l.label){case 0:for(i in n=!1,this.subscriptions)if(i.startsWith(t)){for(n=!0,s=0,a=this.subscriptions[i];s<a.length;s++)u=a[s],(e=this.eventSource)===null||e===void 0||e.removeEventListener(i,u);delete this.subscriptions[i]}return n?this.hasSubscriptionListeners()?[4,this.submitSubscriptions()]:[3,2]:[2];case 1:return l.sent(),[3,3];case 2:this.disconnect(),l.label=3;case 3:return[2]}})})},o.prototype.unsubscribeByTopicAndListener=function(t,e){var n;return m(this,void 0,void 0,function(){var i,s;return b(this,function(a){switch(a.label){case 0:if(!Array.isArray(this.subscriptions[t])||!this.subscriptions[t].length)return[2];for(i=!1,s=this.subscriptions[t].length-1;s>=0;s--)this.subscriptions[t][s]===e&&(i=!0,delete this.subscriptions[t][s],this.subscriptions[t].splice(s,1),(n=this.eventSource)===null||n===void 0||n.removeEventListener(t,e));return i?(this.subscriptions[t].length||delete this.subscriptions[t],this.hasSubscriptionListeners()?[3,1]:(this.disconnect(),[3,3])):[2];case 1:return this.hasSubscriptionListeners(t)?[3,3]:[4,this.submitSubscriptions()];case 2:a.sent(),a.label=3;case 3:return[2]}})})},o.prototype.hasSubscriptionListeners=function(t){var e,n;if(this.subscriptions=this.subscriptions||{},t)return!!(!((e=this.subscriptions[t])===null||e===void 0)&&e.length);for(var i in this.subscriptions)if(!((n=this.subscriptions[i])===null||n===void 0)&&n.length)return!0;return!1},o.prototype.submitSubscriptions=function(){return m(this,void 0,void 0,function(){return b(this,function(t){return this.clientId?(this.addAllSubscriptionListeners(),this.lastSentTopics=this.getNonEmptySubscriptionTopics(),[2,this.client.send("/api/realtime",{method:"POST",body:{clientId:this.clientId,subscriptions:this.lastSentTopics},params:{$cancelKey:this.getSubscriptionsCancelKey()}}).catch(function(e){if(!(e!=null&&e.isAbort))throw e})]):[2]})})},o.prototype.getSubscriptionsCancelKey=function(){return"realtime_"+this.clientId},o.prototype.getNonEmptySubscriptionTopics=function(){var t=[];for(var e in this.subscriptions)this.subscriptions[e].length&&t.push(e);return t},o.prototype.addAllSubscriptionListeners=function(){if(this.eventSource)for(var t in this.removeAllSubscriptionListeners(),this.subscriptions)for(var e=0,n=this.subscriptions[t];e<n.length;e++){var i=n[e];this.eventSource.addEventListener(t,i)}},o.prototype.removeAllSubscriptionListeners=function(){if(this.eventSource)for(var t in this.subscriptions)for(var e=0,n=this.subscriptions[t];e<n.length;e++){var i=n[e];this.eventSource.removeEventListener(t,i)}},o.prototype.connect=function(){return m(this,void 0,void 0,function(){var t=this;return b(this,function(e){return this.reconnectAttempts>0?[2]:[2,new Promise(function(n,i){t.pendingConnects.push({resolve:n,reject:i}),t.pendingConnects.length>1||t.initConnect()})]})})},o.prototype.initConnect=function(){var t=this;this.disconnect(!0),clearTimeout(this.connectTimeoutId),this.connectTimeoutId=setTimeout(function(){t.connectErrorHandler(new Error("EventSource connect took too long."))},this.maxConnectTimeout),this.eventSource=new EventSource(this.client.buildUrl("/api/realtime")),this.eventSource.onerror=function(e){t.connectErrorHandler(new Error("Failed to establish realtime connection."))},this.eventSource.addEventListener("PB_CONNECT",function(e){var n=e;t.clientId=n==null?void 0:n.lastEventId,t.submitSubscriptions().then(function(){return m(t,void 0,void 0,function(){var i;return b(this,function(s){switch(s.label){case 0:i=3,s.label=1;case 1:return this.hasUnsentSubscriptions()&&i>0?(i--,[4,this.submitSubscriptions()]):[3,3];case 2:return s.sent(),[3,1];case 3:return[2]}})})}).then(function(){for(var i=0,s=t.pendingConnects;i<s.length;i++)s[i].resolve();t.pendingConnects=[],t.reconnectAttempts=0,clearTimeout(t.reconnectTimeoutId),clearTimeout(t.connectTimeoutId)}).catch(function(i){t.clientId="",t.connectErrorHandler(i)})})},o.prototype.hasUnsentSubscriptions=function(){var t=this.getNonEmptySubscriptionTopics();if(t.length!=this.lastSentTopics.length)return!0;for(var e=0,n=t;e<n.length;e++){var i=n[e];if(!this.lastSentTopics.includes(i))return!0}return!1},o.prototype.connectErrorHandler=function(t){var e=this;if(clearTimeout(this.connectTimeoutId),clearTimeout(this.reconnectTimeoutId),!this.clientId&&!this.reconnectAttempts||this.reconnectAttempts>this.maxReconnectAttempts){for(var n=0,i=this.pendingConnects;n<i.length;n++)i[n].reject(new O(t));return this.pendingConnects=[],void this.disconnect()}this.disconnect(!0);var s=this.predefinedReconnectIntervals[this.reconnectAttempts]||this.predefinedReconnectIntervals[this.predefinedReconnectIntervals.length-1];this.reconnectAttempts++,this.reconnectTimeoutId=setTimeout(function(){e.initConnect()},s)},o.prototype.disconnect=function(t){var e;if(t===void 0&&(t=!1),clearTimeout(this.connectTimeoutId),clearTimeout(this.reconnectTimeoutId),this.removeAllSubscriptionListeners(),this.client.cancelRequest(this.getSubscriptionsCancelKey()),(e=this.eventSource)===null||e===void 0||e.close(),this.eventSource=null,this.clientId="",!t){this.reconnectAttempts=0;for(var n=0,i=this.pendingConnects;n<i.length;n++)i[n].resolve();this.pendingConnects=[]}},o}(_),ht=function(r){function o(){return r!==null&&r.apply(this,arguments)||this}return y(o,r),o.prototype.check=function(t){return t===void 0&&(t={}),this.client.send("/api/health",{method:"GET",params:t})},o}(_),pt=function(r){function o(){return r!==null&&r.apply(this,arguments)||this}return y(o,r),o.prototype.getUrl=function(t,e,n){n===void 0&&(n={});var i=[];i.push("api"),i.push("files"),i.push(encodeURIComponent(t.collectionId||t.collectionName)),i.push(encodeURIComponent(t.id)),i.push(encodeURIComponent(e));var s=this.client.buildUrl(i.join("/"));if(Object.keys(n).length){var a=new URLSearchParams(n);s+=(s.includes("?")?"&":"?")+a}return s},o.prototype.getToken=function(t){return t===void 0&&(t={}),this.client.send("/api/files/token",{method:"POST",params:t}).then(function(e){return(e==null?void 0:e.token)||""})},o}(_),ft=function(r){function o(){return r!==null&&r.apply(this,arguments)||this}return y(o,r),o.prototype.getFullList=function(t){return t===void 0&&(t={}),this.client.send("/api/backups",{method:"GET",params:t})},o.prototype.create=function(t,e){e===void 0&&(e={});var n={name:t};return this.client.send("/api/backups",{method:"POST",params:e,body:n}).then(function(){return!0})},o.prototype.delete=function(t,e){return e===void 0&&(e={}),this.client.send("/api/backups/".concat(encodeURIComponent(t)),{method:"DELETE",params:e}).then(function(){return!0})},o.prototype.restore=function(t,e){return e===void 0&&(e={}),this.client.send("/api/backups/".concat(encodeURIComponent(t),"/restore"),{method:"POST",params:e}).then(function(){return!0})},o.prototype.getDownloadUrl=function(t,e){return this.client.buildUrl("/api/backups/".concat(encodeURIComponent(e),"?token=").concat(encodeURIComponent(t)))},o}(_),A=function(){function r(o,t,e){o===void 0&&(o="/"),e===void 0&&(e="en-US"),this.cancelControllers={},this.recordServices={},this.enableAutoCancellation=!0,this.baseUrl=o,this.lang=e,this.authStore=t||new nt,this.admins=new it(this),this.collections=new ut(this),this.files=new pt(this),this.logs=new lt(this),this.settings=new ot(this),this.realtime=new dt(this),this.health=new ht(this),this.backups=new ft(this)}return r.prototype.collection=function(o){return this.recordServices[o]||(this.recordServices[o]=new st(this,o)),this.recordServices[o]},r.prototype.autoCancellation=function(o){return this.enableAutoCancellation=!!o,this},r.prototype.cancelRequest=function(o){return this.cancelControllers[o]&&(this.cancelControllers[o].abort(),delete this.cancelControllers[o]),this},r.prototype.cancelAllRequests=function(){for(var o in this.cancelControllers)this.cancelControllers[o].abort();return this.cancelControllers={},this},r.prototype.send=function(o,t){var e,n,i,s,a,u,l,d;return m(this,void 0,void 0,function(){var c,h,f,v,C,g,I,p,T,L=this;return b(this,function(k){switch(k.label){case 0:return c=Object.assign({method:"GET"},t),this.isFormData(c.body)||(c.body&&typeof c.body!="string"&&(c.body=JSON.stringify(c.body)),((e=c==null?void 0:c.headers)===null||e===void 0?void 0:e["Content-Type"])===void 0&&(c.headers=Object.assign({},c.headers,{"Content-Type":"application/json"}))),((n=c==null?void 0:c.headers)===null||n===void 0?void 0:n["Accept-Language"])===void 0&&(c.headers=Object.assign({},c.headers,{"Accept-Language":this.lang})),!((i=this.authStore)===null||i===void 0)&&i.token&&((s=c==null?void 0:c.headers)===null||s===void 0?void 0:s.Authorization)===void 0&&(c.headers=Object.assign({},c.headers,{Authorization:this.authStore.token})),this.enableAutoCancellation&&((a=c.params)===null||a===void 0?void 0:a.$autoCancel)!==!1&&(h=((u=c.params)===null||u===void 0?void 0:u.$cancelKey)||(c.method||"GET")+o,this.cancelRequest(h),f=new AbortController,this.cancelControllers[h]=f,c.signal=f.signal),(l=c.params)===null||l===void 0||delete l.$autoCancel,(d=c.params)===null||d===void 0||delete d.$cancelKey,v=this.buildUrl(o),c.params!==void 0&&((C=this.serializeQueryParams(c.params))&&(v+=(v.includes("?")?"&":"?")+C),delete c.params),this.beforeSend?(p=(I=Object).assign,T=[{}],[4,this.beforeSend(v,c)]):[3,2];case 1:(g=p.apply(I,T.concat([k.sent()]))).url!==void 0||g.options!==void 0?(v=g.url||v,c=g.options||c):Object.keys(g).length&&(c=g,console!=null&&console.warn&&console.warn("Deprecated format of beforeSend return: please use `return { url, options }`, instead of `return options`.")),k.label=2;case 2:return[2,fetch(v,c).then(function(S){return m(L,void 0,void 0,function(){var w;return b(this,function(P){switch(P.label){case 0:w={},P.label=1;case 1:return P.trys.push([1,3,,4]),[4,S.json()];case 2:return w=P.sent(),[3,4];case 3:return P.sent(),[3,4];case 4:return this.afterSend?[4,this.afterSend(S,w)]:[3,6];case 5:w=P.sent(),P.label=6;case 6:if(S.status>=400)throw new O({url:S.url,status:S.status,data:w});return[2,w]}})})}).catch(function(S){throw new O(S)})]}})})},r.prototype.getFileUrl=function(o,t,e){return e===void 0&&(e={}),this.files.getUrl(o,t,e)},r.prototype.buildUrl=function(o){var t,e=this.baseUrl;return typeof window>"u"||!window.location||e.startsWith("https://")||e.startsWith("http://")||(e=!((t=window.location.origin)===null||t===void 0)&&t.endsWith("/")?window.location.origin.substring(0,window.location.origin.length-1):window.location.origin||"",this.baseUrl.startsWith("/")||(e+=window.location.pathname||"/",e+=e.endsWith("/")?"":"/"),e+=this.baseUrl),o&&(e+=e.endsWith("/")?"":"/",e+=o.startsWith("/")?o.substring(1):o),e},r.prototype.isFormData=function(o){return o&&(o.constructor.name==="FormData"||typeof FormData<"u"&&o instanceof FormData)},r.prototype.serializeQueryParams=function(o){var t=[];for(var e in o)if(o[e]!==null){var n=o[e],i=encodeURIComponent(e);if(Array.isArray(n))for(var s=0,a=n;s<a.length;s++){var u=a[s];t.push(i+"="+encodeURIComponent(u))}else n instanceof Date?t.push(i+"="+encodeURIComponent(n.toISOString())):typeof n!==null&&typeof n=="object"?t.push(i+"="+encodeURIComponent(JSON.stringify(n))):t.push(i+"="+encodeURIComponent(n))}return t.join("&")},r}();const E=globalThis.__sveltekit_13hb9vh.env;function St(r){const l=new Date(r).getTime();let c=new Date().getTime()-l,h={year:0,months_over:0,weeks_over:0,days_over:0};return h.year=Math.trunc(c/315576e5),c-=h.year*315576e5,h.months_over=Math.trunc(c/26298e5),c-=h.months_over*26298e5,h.weeks_over=Math.trunc(c/6048e5),c-=h.weeks_over*6048e5,h.days_over=Math.trunc(c/864e5),h}const R=r=>r?JSON.parse(JSON.stringify(r)):null,vt=async()=>{const r=new A(E.PUBLIC_PB_URI);return R(await r.collection("users").getFullList(0,{filter:"role = 0"}))[0]},Ct=async()=>{const r=new A(E.PUBLIC_PB_URI);return R(await r.collection("posts").getList(1,3,{sort:"-created"})).items},Ot=async r=>{await new A(E.PUBLIC_PB_URI).collection("posts").delete(r)},It=async r=>R(await r.collection("users").getFullList(0,{filter:"role = 1",sort:"+first_name, +last_name"})),Pt=async(r,o)=>{const t=new A(E.PUBLIC_PB_URI);try{return{success:!0,chats:R(await t.collection("chats").getFullList(0,{filter:`(from = "${r}" && to = "${o}") || (from = "${o}" && to = "${r}")`,sort:"+created"}))}}catch(e){return console.log(e),{error:!0,message:e.message}}},mt=async r=>{const o=new A(E.PUBLIC_PB_URI);return R(await o.collection("chats").getFullList(0,{filter:`(read = false) && (to = "${r}")`,sort:"+created"}))},_t=async r=>{const o=new A(E.PUBLIC_PB_URI),t=await mt(r);if(t.length!=0)for(let e of t)await o.collection("chats").update(e.id,{read:!0})},At=async r=>{const o=new A(E.PUBLIC_PB_URI),t=await vt(),e=R(await o.collection("chats").getFullList(0,{filter:`(from = "${r}" && to = "${t.id}") && (read = false)`,sort:"+created"}));if(e.length)for(let n of e)await o.collection("chats").update(n.id,{read:!0})};export{A as O,z as a,E as b,vt as c,Ct as d,wt as e,Pt as f,It as g,mt as h,St as i,Ot as j,_t as k,At as u};
