+++ frontend/app/client/_app/immutable/chunks/navigation.488f8e7d.js
@@ -1,0 +1,1 @@
+import{k as o}from"./singletons.b2371634.js";const a=o("goto"),i=o("invalidate_all");export{a as g,i};
