--- frontend/app/client/_app/immutable/chunks/navigation.5a144dea.js
@@ -1,1 +1,0 @@
-import{k as o}from"./singletons.9380a151.js";const a=o("goto"),i=o("invalidate_all");export{a as g,i};
