--- frontend/app/client/_app/immutable/nodes/2.c348ed8e.js
@@ -1,1 +1,0 @@
-import{s as K,H as O,f as i,a as L,g as c,h as g,x as w,d as p,c as N,j as u,v as E,i as T,y as o,I as U,J as V,K as W,M as D}from"../chunks/scheduler.22c66fb6.js";import{S as F,i as G,a as P,t as Q}from"../chunks/index.94eb7729.js";import{g as J}from"../chunks/navigation.5a144dea.js";import{a as R,g as X}from"../chunks/store.565d065b.js";function Y(r){let a,_,t,h,l,m="Work Experiences",y,k,f,S="Education",I,b,d,q="Biography",$,x,v;const A=r[2].default,s=O(A,r,r[1],null);return{c(){a=i("main"),_=i("nav"),t=i("ul"),h=i("li"),l=i("a"),l.textContent=m,y=L(),k=i("li"),f=i("a"),f.textContent=S,I=L(),b=i("li"),d=i("a"),d.textContent=q,$=L(),x=i("section"),s&&s.c(),this.h()},l(e){a=c(e,"MAIN",{class:!0});var n=g(a);_=c(n,"NAV",{class:!0});var M=g(_);t=c(M,"UL",{class:!0});var C=g(t);h=c(C,"LI",{class:!0});var j=g(h);l=c(j,"A",{href:!0,"data-svelte-h":!0}),w(l)!=="svelte-1kr4o0s"&&(l.textContent=m),j.forEach(p),y=N(C),k=c(C,"LI",{class:!0});var z=g(k);f=c(z,"A",{href:!0,"data-svelte-h":!0}),w(f)!=="svelte-11l7ynq"&&(f.textContent=S),z.forEach(p),I=N(C),b=c(C,"LI",{class:!0});var B=g(b);d=c(B,"A",{href:!0,"data-svelte-h":!0}),w(d)!=="svelte-dr1zk7"&&(d.textContent=q),B.forEach(p),C.forEach(p),M.forEach(p),$=N(n),x=c(n,"SECTION",{class:!0});var H=g(x);s&&s.l(H),H.forEach(p),n.forEach(p),this.h()},h(){u(l,"href","/background/workexperience"),E(l,"active",r[0]==="workexperience"),u(h,"class","svelte-l50kx3"),u(f,"href","/background/education"),E(f,"active",r[0]==="education"),u(k,"class","svelte-l50kx3"),u(d,"href","/background/biography"),E(d,"active",r[0]==="biography"),u(b,"class","svelte-l50kx3"),u(t,"class","svelte-l50kx3"),u(_,"class","svelte-l50kx3"),u(x,"class","svelte-l50kx3"),u(a,"class","svelte-l50kx3")},m(e,n){T(e,a,n),o(a,_),o(_,t),o(t,h),o(h,l),o(t,y),o(t,k),o(k,f),o(t,I),o(t,b),o(b,d),o(a,$),o(a,x),s&&s.m(x,null),v=!0},p(e,[n]){(!v||n&1)&&E(l,"active",e[0]==="workexperience"),(!v||n&1)&&E(f,"active",e[0]==="education"),(!v||n&1)&&E(d,"active",e[0]==="biography"),s&&s.p&&(!v||n&2)&&U(s,A,e,e[1],v?W(A,e[1],n,null):V(e[1]),null)},i(e){v||(P(s,e),v=!0)},o(e){Q(s,e),v=!1},d(e){e&&p(a),s&&s.d(e)}}}function Z(r,a,_){let t;D(r,X,m=>_(0,t=m));let{$$slots:h={},$$scope:l}=a;return t?J(`/background/${t}`):J("/background/workexperience"),R.set("background"),r.$$set=m=>{"$$scope"in m&&_(1,l=m.$$scope)},[t,l,h]}class le extends F{constructor(a){super(),G(this,a,Z,Y,K,{})}}export{le as component};
