+++ frontend/app/client/_app/version.json
@@ -1,1 +1,1 @@
-{"version":"1751812712050"}
+{"version":"1751815277961"}
