--- frontend/app/server/chunks/2-5be6e831.js
@@ -1,9 +1,0 @@
-const index = 2;
-let component_cache;
-const component = async () => component_cache ??= (await import('./_layout.svelte-198b9bd9.js')).default;
-const imports = ["_app/immutable/nodes/2.c348ed8e.js","_app/immutable/chunks/scheduler.22c66fb6.js","_app/immutable/chunks/index.94eb7729.js","_app/immutable/chunks/navigation.5a144dea.js","_app/immutable/chunks/singletons.9380a151.js","_app/immutable/chunks/index.80348aca.js","_app/immutable/chunks/store.565d065b.js"];
-const stylesheets = ["_app/immutable/assets/2.6f65f4a3.css"];
-const fonts = [];
-
-export { component, fonts, imports, index, stylesheets };
-//# sourceMappingURL=2-5be6e831.js.map
