+++ frontend/app/server/index.js
@@ -198,7 +198,7 @@ const options = {
 		<div class="error">
 			<span class="status">` + status + '</span>\n			<div class="message">\n				<h1>' + message + "</h1>\n			</div>\n		</div>\n	</body>\n</html>\n"
   },
-  version_hash: "13hb9vh"
+  version_hash: "1a5tsyf"
 };
 function get_hooks() {
   return import('./chunks/hooks.server-421e8d93.js');
